Move API client into its own folder
